fix for repayment info
